GFW OGL4 + Implementation of render targets

+++ gfw/source/ogl4/device.cpp
@@ -7,12 +7,12 @@
 #include "gfw/core/context.h"
 #include "gfw/core/device.h"
 #include "gfw/core/drawing_context.h"
+#include "gfw/core/functions.h"
 #include "gfw/core/input_layout.h"
+#include "gfw/core/render_target.h"
 #include "gfw/core/shader.h"
 #include "gfw/core/shader_reflect.h"
 #include "gfw/core/texture.h"
-
-#include "gfw/core/functions.h"
 
 #define AUTO_LOCK_CONTEXT   AutoLock __auto_lock_context(mDrawingContext.GetPointer(), &mMutex, mContextGL)
 
@@ -65,15 +65,21 @@ namespace GFW {
         : mParams(params)
         , mContextGL(NULL)
     {
-
+        TRACE_ASSERT(mParams.backBufferFormat != FORMAT_UNKNOWN);
     }
 
     Device::~Device()
     {
-        TRACE_ASSERT(mDrawingContext.IsAttached());
-        TRACE_ASSERT(mContextGL != NULL);
-        mDrawingContext->MakeCurrent(NULL);
-        mDrawingContext->DeleteContext(mContextGL);
+        if (mResolveFramebuffer != 0)
+        {
+            TRACE_ASSERT_GL(glDeleteFramebuffers, 1, &mResolveFramebuffer);
+        }
+
+        if (mDrawingContext.IsAttached() && mContextGL != NULL)
+        {
+            mDrawingContext->MakeCurrent(NULL);
+            mDrawingContext->DeleteContext(mContextGL);
+        }
     }
 
     bool Device::Initialize()
@@ -94,11 +100,36 @@ namespace GFW {
 
         AUTO_LOCK_CONTEXT;
 
+        InitializeSwapChain();
+
         mDefaultContext = new Context(mDrawingContext, this);
 
         const uint8_t * extensions = glGetString(GL_EXTENSIONS);
 
         return true;
+    }
+
+    void Device::InitializeSwapChain()
+    {
+        TextureDesc defaultRenderTargetTextureDesc;
+        defaultRenderTargetTextureDesc.format       = mParams.backBufferFormat;
+        defaultRenderTargetTextureDesc.width        = mParams.backBufferWidth;
+        defaultRenderTargetTextureDesc.height       = mParams.backBufferHeight;
+        defaultRenderTargetTextureDesc.usage        = USAGE_DEFAULT;
+        defaultRenderTargetTextureDesc.mipLevels    = 1;
+        TextureRef defaultRenderTargetTexture = new Texture(defaultRenderTargetTextureDesc, NULL, this);
+
+        RenderTargetDesc defaultRenderTargetDesc;
+        defaultRenderTargetDesc.format = mParams.backBufferFormat;
+        mDefaultRenderTarget = new RenderTarget(defaultRenderTargetTexture, defaultRenderTargetDesc, this);
+
+        TRACE_ASSERT_GL(glGenFramebuffers, 1, &mResolveFramebuffer);
+        TRACE_ASSERT_GL(glBindFramebuffer, GL_READ_FRAMEBUFFER, mResolveFramebuffer);
+        TRACE_ASSERT_GL(glFramebufferTexture, GL_READ_FRAMEBUFFER, GL_COLOR_ATTACHMENT0, defaultRenderTargetTexture->GetHandle(), 0);
+        TRACE_ASSERT_GL(glReadBuffer, GL_COLOR_ATTACHMENT0);
+
+        TRACE_ASSERT_GL(glBindFramebuffer, GL_DRAW_FRAMEBUFFER, 0);
+        TRACE_ASSERT_GL(glDrawBuffer, GL_LEFT);
     }
 
     IContextRef Device::CreateContext()
@@ -145,17 +176,19 @@ namespace GFW {
         return new Texture(desc, initialData, this);
     }
 
-    IRenderBufferRef Device::CreateRenderBuffer( ITextureIn, const SubResIdx & )
+    IRenderTargetRef Device::CreateRenderTarget(ITextureIn texture, const RenderTargetDesc & desc)
     {
         AUTO_LOCK_CONTEXT;
 
-        TRACE_FAIL_MSG("Not yet implemented");
-        return NULL;
+        return new RenderTarget(texture, desc, this);
     }
 
     void Device::Present()
     {
         AUTO_LOCK_CONTEXT;
+
+        TRACE_ASSERT_GL(glBlitFramebuffer, 0, 0, mParams.backBufferWidth, mParams.backBufferHeight,
+            0, 0, mParams.backBufferWidth, mParams.backBufferHeight, GL_COLOR_BUFFER_BIT, GL_NEAREST);
 
         mDrawingContext->SwapBuffers();
     }
